requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.1
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-nested-routers==0.94.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
flower==2.0.1
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
locust==2.32.5
MarkupSafe==3.0.2
model-bakery==1.20.0
msgpack==1.1.0
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psutil==6.1.1
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
pyzmq==26.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.6.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tornado==6.4.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
